Rename module back to pyap, project name pyap2
