+++ pyap/utils.py
