+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "argyle-pyap"
+name = "pyap2"
 version = "0.0.0"
-description = "Argyle's Pyap is a fork of Pyap, a MIT Licensed text processing library, written in Python, for detecting and parsing addresses. Currently it supports USA, Canadian and British addresses."
+description = "Pyap2 is a maintained fork of pyap, a regex-based library for parsing US, CA, and UK addresses. The fork adds typing support, handling of various address formats and edge cases."
 authors = ["Argyle Developers <developers@argyle.io>"]
 documentation = "https://github.com/argyle-engineering/pyap"
 license = "MIT"
